requirements.txt
openai-agents>=0.0.10
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
python-dotenv>=1.0.0
aiosqlite>=0.21.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
